extract StringRepresentable from RickAndMortyCharacterStatus for future reuse
